requirements.txt
ipdb==0.13.9
pandas==1.3.4
psycopg2_binary==2.9.1
sqlalchemy==1.4.22